SabrePatrolLPR/mdt_client/requirements.txt
PyQt5
opencv-python-headless
requests
websocket-client
pyinstaller
